Cleaned up binding callback observer interface and c_api
