+++ Package.swift
@@ -93,7 +93,6 @@ let notSyncServerSources: [String] = [
     "realm/spec.cpp",
     "realm/status.cpp",
     "realm/string_data.cpp",
-    "realm/sync/binding_callback_thread_observer.cpp",
     "realm/sync/changeset.cpp",
     "realm/sync/changeset_encoder.cpp",
     "realm/sync/changeset_parser.cpp",
