+++ src/realm.h
@@ -3474,6 +3474,12 @@ RLM_API void realm_sync_client_config_set_pong_keepalive_timeout(realm_sync_clie
                                                                  uint64_t) RLM_API_NOEXCEPT;
 RLM_API void realm_sync_client_config_set_fast_reconnect_limit(realm_sync_client_config_t*,
                                                                uint64_t) RLM_API_NOEXCEPT;
+RLM_API void realm_sync_client_config_set_sync_socket(realm_sync_client_config_t*,
+                                                      realm_sync_socket_t*) RLM_API_NOEXCEPT;
+RLM_API void realm_sync_client_config_set_default_binding_thread_observer(
+    realm_sync_client_config_t* config, realm_on_object_store_thread_callback_t on_thread_create,
+    realm_on_object_store_thread_callback_t on_thread_destroy, realm_on_object_store_error_callback_t on_error,
+    realm_userdata_t user_data, realm_free_userdata_func_t free_userdata);
 
 RLM_API realm_sync_config_t* realm_sync_config_new(const realm_user_t*, const char* partition_value) RLM_API_NOEXCEPT;
 RLM_API realm_sync_config_t* realm_flx_sync_config_new(const realm_user_t*) RLM_API_NOEXCEPT;
@@ -3845,27 +3851,6 @@ RLM_API void realm_sync_session_handle_error_for_testing(const realm_sync_sessio
  */
 RLM_API void realm_register_user_code_callback_error(realm_userdata_t usercode_error) RLM_API_NOEXCEPT;
 
-#if REALM_ENABLE_SYNC
-
-typedef struct realm_thread_observer_token realm_thread_observer_token_t;
-
-/**
- * Register an app local callback handler for bindings interested in registering callbacks before/after
- * the ObjectStore thread runs for this app. This only works for the default socket provider implementation.
- * @param config SyncClientConfig ptr created by realm_sync_client_config_new()
- * @param on_thread_create callback invoked when the object store thread is created
- * @param on_thread_destroy callback invoked when the object store thread is destroyed
- * @param on_error callback invoked to signal to the listener that some error has occured.
- * @param user_data pointer to user defined data that is provided to each of the callback functions
- * @param free_userdata callback invoked when the user_data is to be freed
- */
-RLM_API void realm_sync_client_config_set_default_binding_thread_observer(
-    realm_sync_client_config_t* config, realm_on_object_store_thread_callback_t on_thread_create,
-    realm_on_object_store_thread_callback_t on_thread_destroy, realm_on_object_store_error_callback_t on_error,
-    realm_userdata_t user_data, realm_free_userdata_func_t free_userdata);
-
-#endif // REALM_ENABLE_SYNC
-
 typedef struct realm_mongodb_collection realm_mongodb_collection_t;
 
 typedef struct realm_mongodb_find_options {
@@ -4105,8 +4090,5 @@ RLM_API void realm_sync_socket_websocket_message(realm_websocket_observer_t* rea
 
 RLM_API void realm_sync_socket_websocket_closed(realm_websocket_observer_t* realm_websocket_observer, bool was_clean,
                                                 realm_web_socket_errno_e status, const char* reason);
-
-RLM_API void realm_sync_client_config_set_sync_socket(realm_sync_client_config_t*,
-                                                      realm_sync_socket_t*) RLM_API_NOEXCEPT;
 
 #endif // REALM_H
