+++ src/realm/object-store/c_api/realm.hpp
@@ -21,10 +21,6 @@
 #include <realm/object-store/c_api/util.hpp>
 #include <realm/object-store/binding_context.hpp>
 
-#if REALM_ENABLE_SYNC
-#include <realm/sync/binding_callback_thread_observer.hpp>
-#endif // REALM_ENABLE_SYNC
-
 namespace realm::c_api {
 
 class CBindingContext : public BindingContext {
@@ -66,74 +62,5 @@ private:
     CallbackRegistryWithVersion<> m_realm_pending_refresh_callbacks;
     CallbackRegistry<const Schema&> m_schema_changed_callbacks;
 };
-
-#if REALM_ENABLE_SYNC
-
-struct CBindingThreadObserver : public realm::BindingCallbackThreadObserver {
-public:
-    CBindingThreadObserver(realm_on_object_store_thread_callback_t on_thread_create,
-                           realm_on_object_store_thread_callback_t on_thread_destroy,
-                           realm_on_object_store_error_callback_t on_error, realm_userdata_t userdata,
-                           realm_free_userdata_func_t free_userdata)
-        : m_create_callback_func{on_thread_create}
-        , m_destroy_callback_func{on_thread_destroy}
-        , m_error_callback_func{on_error}
-        , m_user_data{UserdataPtr(userdata, free_userdata)}
-    {
-    }
-
-    virtual ~CBindingThreadObserver() = default;
-
-    /// {@
-    /// For testing: Return the values in this CBindingThreadObserver for comparing if two objects
-    /// have the same callback functions and userdata ptr values.
-    inline realm_on_object_store_thread_callback_t get_create_callback_func() const noexcept
-    {
-        return m_create_callback_func;
-    }
-    inline realm_on_object_store_thread_callback_t get_destroy_callback_func() const noexcept
-    {
-        return m_destroy_callback_func;
-    }
-    inline realm_on_object_store_error_callback_t get_error_callback_func() const noexcept
-    {
-        return m_error_callback_func;
-    }
-    inline realm_userdata_t get_userdata_ptr() const noexcept
-    {
-        return m_user_data.get();
-    }
-    /// @}
-
-protected:
-    CBindingThreadObserver() = default;
-
-    void did_create_thread() override
-    {
-        if (m_create_callback_func)
-            m_create_callback_func(m_user_data.get());
-    }
-
-    void will_destroy_thread() override
-    {
-        if (m_destroy_callback_func)
-            m_destroy_callback_func(m_user_data.get());
-    }
-
-    bool handle_error(std::exception const& e) override
-    {
-        if (!m_error_callback_func)
-            return false;
-
-        return m_error_callback_func(m_user_data.get(), e.what());
-    }
-
-    realm_on_object_store_thread_callback_t m_create_callback_func = nullptr;
-    realm_on_object_store_thread_callback_t m_destroy_callback_func = nullptr;
-    realm_on_object_store_error_callback_t m_error_callback_func = nullptr;
-    UserdataPtr m_user_data;
-};
-
-#endif // REALM_ENABLE_SYNC
 
 } // namespace realm::c_api
